LayoutTests/fast/forms/input-step-unsupported-expected.txt

Tests stepDown()/stepUp() for unsupported types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.stepDown() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.stepDown(0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.stepUp() threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS input.stepUp(0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
